animate-dead: cap scaled stats (cap=0 = no cap; multiplier cap for HP)
Add a bound to the per-construct <scaling> stats:
- cap="0" or an absent cap now means NO cap (was: a real cap of 0).
- The cap field is a double; for mode=mult (HP) it bounds the MULTIPLIER
  (hp <= proto*cap), clean since prototype HP is a dice roll. For mode=add
  it stays an absolute ceiling (up) / floor (down: ac, saving).
- Data: HP cap=2.0 (~ the old remort-10 ceiling) and damage_dice cap=25 on
  all tiers, bounding the competence scaling near the pre-rework maximum.
  Saving-throw scaling stays off (beta=0), so its floor cap is a no-op for now.

+++ src/gameplay/mechanics/animate_dead.h
@@ -26,8 +26,8 @@ enum class EScaleMode { kAdd, kMult };
 struct StatScale {
 	double beta = 0.0;
 	EScaleMode mode = EScaleMode::kAdd;
-	bool has_cap = false;
-	int cap = 0;
+	bool has_cap = false;   // set from cap!=0. mode=mult: cap is the max MULTIPLIER (hp <= proto*cap);
+	double cap = 0;         // mode=add "up": ceiling; "down" (ac/saving): floor. cap="0"/absent = no cap.
 };
 
 // Per-tier scaling. 'luck' is intentionally absent (mobs have no luck stat).
